Add efficient check of hamming weight difference upon subtraction
This new function, hamming_weight_change_upon_subtraction, gives the
difference between hamming weight of x and (x-y). This is done in
O(y.hamming_weight()), which is useful if y.hamming_weight is much
smaller than y.len() or x.len(). This is indeed the case in the
part of the decryption routine that sorts by hamming weight.